remove int type errors
the UINT4 type musst be redefined because its definition is not a 32 bit word today. Otherwise the program calcualtes wrong hash words and the selftest (option -x) will fail.

+++ global.h
@@ -1,6 +1,8 @@
 /* GLOBAL.H - RSAREF types and constants
  */
 
+#include <inttypes.h>
+
 /* PROTOTYPES should be set to one if and only if the compiler supports
   function argument prototyping.
 The following makes PROTOTYPES default to 0 if it has not already
@@ -14,10 +16,10 @@ The following makes PROTOTYPES default to 0 if it has not already
 typedef unsigned char *POINTER;
 
 /* UINT2 defines a two byte word */
-typedef unsigned short int UINT2;
+typedef uint16_t UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 
 /* PROTO_LIST is defined depending on how PROTOTYPES is defined above.
 If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
